增加了脚本的错误检测

+++ gitbook.sh
@@ -1,14 +1,21 @@
 #bash
 
-git checkout master
-rm -rf _book
-gitbook build
-git add .
-git commit -m $1
-git push -u origin master
-git checkout gh-pages
-cp -r _book/* .
-git add .
-git commit -m $1
-git push -u origin gh-pages
-git checkout master
+MESSAGE="$1"
+
+if [[ $MESSAGE == "" ]]; then
+  echo "err: 缺少commit的描述信息。"
+else
+  git checkout master
+  rm -rf _book
+  gitbook build
+  git add .
+  git commit -m $MESSAGE
+  git push -u origin master
+  git checkout gh-pages
+  cp -r _book/* .
+  git add .
+  git commit -m $MESSAGE
+  git push -u origin gh-pages
+  git checkout master
+fi
+
